[Streaming] Move resource-manager and scheduler to master package. (#7582)
