+++ streaming/java/streaming-runtime/src/main/java/org/ray/streaming/runtime/master/scheduler/strategy/SlotAssignStrategyFactory.java
@@ -1,7 +1,7 @@
-package org.ray.streaming.runtime.core.master.scheduler.strategy;
+package org.ray.streaming.runtime.master.scheduler.strategy;
 
 import org.ray.streaming.runtime.config.types.SlotAssignStrategyType;
-import org.ray.streaming.runtime.core.master.scheduler.strategy.impl.PipelineFirstStrategy;
+import org.ray.streaming.runtime.master.scheduler.strategy.impl.PipelineFirstStrategy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
