+++ streaming/java/streaming-runtime/src/test/java/org/ray/streaming/runtime/resourcemanager/ResourceManagerTest.java
@@ -13,10 +13,10 @@ import org.ray.streaming.runtime.config.StreamingConfig;
 import org.ray.streaming.runtime.config.global.CommonConfig;
 import org.ray.streaming.runtime.config.master.ResourceConfig;
 import org.ray.streaming.runtime.core.graph.executiongraph.ExecutionGraph;
-import org.ray.streaming.runtime.core.master.resourcemanager.ResourceManager;
-import org.ray.streaming.runtime.core.master.resourcemanager.ResourceManagerImpl;
-import org.ray.streaming.runtime.core.master.scheduler.strategy.SlotAssignStrategy;
-import org.ray.streaming.runtime.core.master.scheduler.strategy.impl.PipelineFirstStrategy;
+import org.ray.streaming.runtime.master.resourcemanager.ResourceManager;
+import org.ray.streaming.runtime.master.resourcemanager.ResourceManagerImpl;
+import org.ray.streaming.runtime.master.scheduler.strategy.SlotAssignStrategy;
+import org.ray.streaming.runtime.master.scheduler.strategy.impl.PipelineFirstStrategy;
 import org.ray.streaming.runtime.core.resource.Container;
 import org.ray.streaming.runtime.core.resource.ContainerID;
 import org.ray.streaming.runtime.core.resource.Slot;
