+++ streaming/java/streaming-runtime/src/test/java/org/ray/streaming/runtime/schedule/strategy/PipelineFirstStrategyTest.java
@@ -7,27 +7,15 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 
-import com.google.common.collect.Lists;
 import org.aeonbits.owner.ConfigFactory;
-import org.ray.api.RayActor;
-import org.ray.api.id.ActorId;
-import org.ray.api.id.ObjectId;
 import org.ray.api.id.UniqueId;
-import org.ray.runtime.actor.LocalModeRayActor;
-import org.ray.streaming.api.context.RuntimeContext;
-import org.ray.streaming.api.context.StreamingContext;
-import org.ray.streaming.api.stream.DataStream;
-import org.ray.streaming.api.stream.DataStreamSink;
-import org.ray.streaming.api.stream.DataStreamSource;
 import org.ray.streaming.jobgraph.JobGraph;
-import org.ray.streaming.jobgraph.JobGraphBuilder;
 import org.ray.streaming.runtime.BaseUnitTest;
 import org.ray.streaming.runtime.config.StreamingConfig;
-import org.ray.streaming.runtime.config.StreamingMasterConfig;
 import org.ray.streaming.runtime.config.master.ResourceConfig;
 import org.ray.streaming.runtime.core.graph.executiongraph.ExecutionGraph;
-import org.ray.streaming.runtime.core.master.scheduler.strategy.SlotAssignStrategy;
-import org.ray.streaming.runtime.core.master.scheduler.strategy.impl.PipelineFirstStrategy;
+import org.ray.streaming.runtime.master.scheduler.strategy.SlotAssignStrategy;
+import org.ray.streaming.runtime.master.scheduler.strategy.impl.PipelineFirstStrategy;
 import org.ray.streaming.runtime.core.resource.Container;
 import org.ray.streaming.runtime.core.resource.ContainerID;
 import org.ray.streaming.runtime.core.resource.Resources;
